requirements.txt
langchain-classic
langchain-community
langchain-groq
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf
sentence-transformers
streamlit
python-dotenv